route protecting bug
