+++ app/pages/note/[id].vue
@@ -5,13 +5,8 @@
 </template>
 <script setup>
 const route = useRoute();
-const authStore = useAuthStore();
 const id = route.params.id;
-onMounted(async () => {
-  const token = useCookie("auth-token").value;
-  const ok = await authStore.checkAuth();
-  if (!token || !ok) {
-    return navigateTo("/login");
-  }
+definePageMeta({
+  middleware: "protected",
 });
 </script>
