+++ app/pages/note/create.vue
@@ -9,11 +9,7 @@ const authStore = useAuthStore();
 const router = useRouter();
 const snackbar = useSnackbarStore();
 
-onMounted(async () => {
-  const token = useCookie("auth-token").value;
-  const ok = await authStore.checkAuth();
-  if (!token || !ok) {
-    return navigateTo("/login");
-  }
+definePageMeta({
+  middleware: "protected",
 });
 </script>
